+++ src/extension.ts
@@ -1,7 +1,6 @@
 import * as vscode from 'vscode';
-import { command, COMMANDS, deleteConflictingOutputsSuffix, isLinux, log, pubCommand, settings } from './utils';
+import { command, COMMANDS, DartFlutterCommand, deleteConflictingOutputsSuffix, isLinux, log, pubCommand, settings } from './utils';
 import { BuildRunnerWatch } from './watch';
-import { DartFlutterCommand } from './utils';
 import p = require('path');
 const yaml = require('js-yaml');
 import fs = require('fs');
@@ -274,7 +273,7 @@ export function getDartProjectPath(): string | undefined {
 	if (fs.existsSync(workspacePath! + pubspecSuffix)) { return workspacePath; }
 
 	const walkSegments: string[] = [...segments];
-	for (let i = segments.length - 1; i >= 0; i--) {
+	for (let i = walkSegments.length; i >= 0; i--) {
 		const projectPath = vscode.Uri.file(p.join(workspacePath, ...walkSegments));
 		const pubspec = vscode.Uri.joinPath(projectPath, pubspecSuffix);
 		console.log('Looking for ' + pubspec.fsPath);