+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to the "build-runner" extension will be documented in this file.
 
+
+## [0.4.0]
+
+- Improved watch on Windows. Now almost fully operational.
+- Fixed issue with finding the project
+
 ## [0.3.7]
 
 - Find pubspec in subfolder first ([#5](https://github.com/gaetschwartz/build-runner/pull/5))